+++ cg/exc.py
@@ -256,7 +256,3 @@ class ArchiveJobFailedError(CgError):
 
 class XMLError(CgError):
     """Exception raised when something is wrong with the content of an XML file."""
-
-
-class OrderNotFoundError(CgError):
-    """Exception raised when an order is not found."""
