+++ cg/services/order_service/exceptions.py
@@ -1,0 +1,6 @@
+class OrderServiceError(Exception):
+    pass
+
+
+class OrderNotFoundError(OrderServiceError):
+    pass
